% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuedfraction.R
\name{continuedfraction}
\alias{continuedfraction}
\title{Return a Vector representing the Continued Fraction for the Input Number}
\usage{
continuedfraction(x, n = 10)
}
\arguments{
\item{x}{Input Number}

\item{n}{Desired Precision (number of fractions)}
}
\value{
Continued Fraction Vector
}
\description{
Return a Vector representing the Continued Fraction for the Input Number
}
\examples{
continuedfraction(1.24) # Do not use decimals when calling the function
continuedfraction(gmp::as.bigq(124,100)) # Use Fractions instead
continuedfraction(sqrt(2))
}
