% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuedfractioninverse.R
\name{continuedfractioninverse}
\alias{continuedfractioninverse}
\title{Inverse Continued Fraction for Simple Continued Fractions}
\usage{
continuedfractioninverse(v)
}
\arguments{
\item{v}{Vector representing the Continued Fraction}
}
\value{
A Number
}
\description{
Inverse Continued Fraction for Simple Continued Fractions
}
\examples{
continuedfractioninverse(c(1,2,2,2,2,2,2,2,2))
}
