% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decimalfixed.R
\name{decimalfixed}
\alias{decimalfixed}
\title{Computes the fraction that originates the given fixed decimal}
\usage{
decimalfixed(fixed)
}
\arguments{
\item{fixed}{A Numeric with or without decimal point}
}
\description{
Computes the fraction that originates the given fixed decimal
}
\examples{
decimalfixed(0.352)
}
