% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decimalrepeating.R
\name{decimalrepeating}
\alias{decimalrepeating}
\title{Computes the fraction that originates the given repeating decimal}
\usage{
decimalrepeating(dec)
}
\arguments{
\item{dec}{A string, with the repeating decimal between square brackets}
}
\value{
fraction
}
\description{
Computes the fraction that originates the given repeating decimal
#############################
\subsection{Author: Roman Chokler}{

#############################
}
}
\examples{
decimalrepeating("1.8[3]")
}
