% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decimalrepeatinginverse.R
\name{decimalrepeatinginverse}
\alias{decimalrepeatinginverse}
\title{Computes the decimal that originates the given fraction}
\usage{
decimalrepeatinginverse(q)
}
\arguments{
\item{q}{gmp::as.bigq(numerator,denominator)}
}
\value{
dec string
}
\description{
Computes the decimal that originates the given fraction
#############################
\subsection{Author: Roman Chokler}{

#############################
}
}
\examples{
decimalrepeatinginverse(gmp::as.bigq(4,7))
}
