% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derangement.R
\name{derangement}
\alias{derangement}
\title{Return the number of Derangements}
\usage{
derangement(n)
}
\arguments{
\item{n}{Input Integer}
}
\value{
Number of Derangements
}
\description{
Return the number of Derangements
}
\examples{
for ( i in 1:10 ) print(derangement(i))
}
