% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derangements.R
\name{derangements}
\alias{derangements}
\title{Return the Data Frame with the Derangements built from the input vector}
\usage{
derangements(v)
}
\arguments{
\item{v}{Input Vector}
}
\value{
derangements dataframe
}
\description{
Return the Data Frame with the Derangements built from the input vector
}
\examples{
derangements(c("A","B","C"))
}
