% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digitalroot.R
\name{digitalroot}
\alias{digitalroot}
\title{Computes the Digital Root of integer/string n}
\usage{
digitalroot(n)
}
\arguments{
\item{n}{Input}
}
\value{
Digital Root
}
\description{
Computes the Digital Root of integer/string n
}
\examples{
digitalroot(123456)
}
