% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editvariable.R
\name{editvariable}
\alias{editvariable}
\title{Allows for editing a string variable}
\usage{
editvariable(x)
}
\arguments{
\item{x}{String variable}
}
\description{
Allows for editing a string variable
}
\examples{
\dontrun{
editvariable(x)
}
}
