% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidnumbers.R
\name{euclidnumbers}
\alias{euclidnumbers}
\title{Generate Euclid Numbers}
\usage{
euclidnumbers(n)
}
\arguments{
\item{n}{nth Euclid Number}
}
\value{
Euclid Number
}
\description{
An Euclid Number is the Product of n first prime numbers, plus 1
}
\examples{
for ( i in 1:6 ) print( euclidnumbers(i) )
}
