% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibonaccisequence.R
\name{fibonaccisequence}
\alias{fibonaccisequence}
\title{Return a vector of fibonnaci numbers}
\usage{
fibonaccisequence(v, n)
}
\arguments{
\item{v}{a vector with the first 2 or more elements of a fibonacci sequence}

\item{n}{the length of the desired sequence}
}
\value{
vector of fibonnaci numbers
}
\description{
Return a vector of fibonnaci numbers
}
\examples{
fibonaccisequence(c(1,1),10)
fibonaccisequence(c(2,3),10)
fibonaccisequence(c(1,2,4),10)
}
