% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfunctionpolynom.R
\name{getfunctionpolynom}
\alias{getfunctionpolynom}
\title{Computes a Polynomial Function ( a formula )
able to generate from the X values received, the correspondent Y values received}
\usage{
getfunctionpolynom(X, Y)
}
\arguments{
\item{X}{A vector of numeric inputs}

\item{Y}{A vector of numeric outputs that must be generated by the returned function}
}
\value{
A function able to generate from the X values received, the correspondent Y values received
}
\description{
Computes a Generating Function able to generate the expected Outputs
}
\examples{
fp = minifunctions::getfunctionpolynom(1:3,c(3,7,21))
for ( i in 1:10 ) print(fp(i))
}
