% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmonths.R
\name{getmonths}
\alias{getmonths}
\title{Return the abbreviations of the 12 months of the year in the current language}
\usage{
getmonths()
}
\value{
months
}
\description{
Return the abbreviations of the 12 months of the year in the current language
}
\examples{
getmonths()
}
