% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highprecision.R
\name{highprecision}
\alias{highprecision}
\title{Operates 2 arguments in High Precision}
\usage{
highprecision(a, operator, b)
}
\arguments{
\item{a}{Argument 1}

\item{operator}{Some operator like "+", "*", etc}

\item{b}{Argument 2}
}
\value{
The Result of the Operation
}
\description{
Operates 2 arguments in High Precision
}
\examples{
highprecision(1,"/",31)
}
