% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minmax.R
\name{minmax}
\alias{minmax}
\title{Return the Min–Max normalization (0–1) for all numeric columns of Data Frame}
\usage{
minmax(df)
}
\arguments{
\item{df}{Input DataFrame}
}
\value{
Normalized DataFrame
}
\description{
Return the Min–Max normalization (0–1) for all numeric columns of Data Frame
}
