% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplicativeorder.R
\name{multiplicativeorder}
\alias{multiplicativeorder}
\title{Compute the Multiplicative Order of A modulo N}
\usage{
multiplicativeorder(a, n)
}
\arguments{
\item{a}{Integer}

\item{n}{Integer Modulo}
}
\value{
Multiplicative Order
}
\description{
In number theory, given an integer A and a positive integer N with GCD(A,N) = 1,
the multiplicative order of A modulo N is the smallest positive integer k with A^k( mod N ) = 1. ( 0 < K < N )
}
\examples{
multiplicativeorder(13,100) # Returns 20
}
