% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newtonmethod.R
\name{newtonmethod}
\alias{newtonmethod}
\title{Find Root using Newton Method}
\usage{
newtonmethod(f, x0, n)
}
\arguments{
\item{f}{A Function}

\item{x0}{Initial guess}

\item{n}{Number of iterations}
}
\value{
Root aproximation for given function
}
\description{
Find Root using Newton Method starting with x0 aproximation and executing n iterations
Newton's method doesn't always work.
Sometimes, things depend on making a good choice for the initial guess
}
\examples{
newtonmethod(function(x) x^2 -3,3,5) # Square Root of 3
}
