% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numberconversionfrac.R
\name{numberconversionfrac}
\alias{numberconversionfrac}
\title{Convert Integer from Base 10 to Fractional Base}
\usage{
numberconversionfrac(n, numerator, denominator)
}
\arguments{
\item{n}{Integer to convert}

\item{numerator}{The numerator of the base}

\item{denominator}{The denominator of the base}
}
\value{
A vector with the numeric representation on the new base
}
\description{
Convert Integer from Base 10 to Fractional Base
}
\examples{
for ( i in 1:20 ) cat(i," = ",numberconversionfrac(i,4,3),"\n")
}
