% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numberconversion.R
\name{numberconvertion}
\alias{numberconvertion}
\title{Convert Integer from Base1 to Base2}
\usage{
numberconvertion(n, base1, base2)
}
\arguments{
\item{n}{Number to convert}

\item{base1}{Input Numeric Base System like 2, 3, 10, 16, etc}

\item{base2}{Output Numeric Base System like 2, 3, 10, 16, etc}
}
\value{
The converted Integer in the new Base
}
\description{
Convert Integer from Base1 to Base2
}
\examples{
n = 100
for ( base in 2:16 ) message(n," ( base 10 ) = ",numberconvertion(n,10,base)," ( base ",base," )")
}
