% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numberoperation.R
\name{numberoperation}
\alias{numberoperation}
\title{Operates 2 arguments in the indicated numeric Base}
\usage{
numberoperation(a, operator, b, base)
}
\arguments{
\item{a}{Argument 1}

\item{operator}{Some operator like "+", "*", etc}

\item{b}{Argument 2}

\item{base}{Numeric Base System like 2, 3, 10, 16, etc}
}
\value{
The Result of the Operation expressed in the indicated Base
}
\description{
Operates 2 arguments in the indicated numeric Base
}
\examples{
numberoperation("222","+","111",3)
}
