% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parabola.R
\name{parabola}
\alias{parabola}
\title{Solve Parabola given Parameters a, b, c}
\usage{
parabola(a, b, c)
}
\arguments{
\item{a}{ax^2 + bx + c}

\item{b}{ax^2 + bx + c}

\item{c}{ax^2 + bx + c}
}
\description{
Solve Parabola given Parameters a, b, c
}
\examples{
parabola(5,-30,49)
}
