% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parabolafromfocus.R
\name{parabolafromfocus}
\alias{parabolafromfocus}
\title{Solve Parabola given Focus(h,f) and Directrix y=d}
\usage{
parabolafromfocus(h, f, d)
}
\arguments{
\item{h}{Focus(h,f)}

\item{f}{Focus(h,f)}

\item{d}{Directrix y=d}
}
\description{
Solve Parabola given Focus(h,f) and Directrix y=d
}
\examples{
parabolafromfocus(18,24,34)
}
