% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutations.R
\name{permutations}
\alias{permutations}
\title{Computes the permutations of vector, allowing for repeated values}
\usage{
permutations(v, L)
}
\arguments{
\item{v}{Input Vector}

\item{L}{Length of the intended permutations ( defaults to the length of v )}
}
\value{
Permutations Dataframe
}
\description{
Computes the permutations of vector, allowing for repeated values
}
\examples{
permutations(c(1,3,4,4))
permutations(c(1,3,4,4),3)
}
