% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerset.R
\name{powerset}
\alias{powerset}
\title{Generate the Powerset of a Vector}
\usage{
powerset(v)
}
\arguments{
\item{v}{Vector}
}
\value{
A List of Vectors
}
\description{
Generate the Powerset of a Vector
}
\examples{
powerset(c("A","B","C"))
}
