% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printdataframe.R
\name{printdataframe}
\alias{printdataframe}
\title{Print Data Frame as HTML Table}
\usage{
printdataframe(df, expand_images = FALSE)
}
\arguments{
\item{df}{Data Frame to be printed as HTML Table}

\item{expand_images}{Boolean: if TRUE, image URLs will be expanded to show the images in the table.}
}
\description{
Print Data Frame as HTML Table
}
\examples{
\dontrun{
printdataframe(iris)
}
}
