% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repunit.R
\name{repunit}
\alias{repunit}
\title{Generate the repunit 1111... of Length n}
\usage{
repunit(n)
}
\arguments{
\item{n}{Length of the Repunit}
}
\value{
repunit
}
\description{
Generate the repunit 1111... of Length n
}
\examples{
repunit(5)
}
