% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serpapi.R
\name{serpapi}
\alias{serpapi}
\title{Submit a google search with SerpApi}
\usage{
serpapi(prompt)
}
\arguments{
\item{prompt}{Text prompt}
}
\description{
Submit a google search with SerpApi
}
\examples{
\dontrun{
serpapi("Qual a capital de Portugal?")
}
}
