% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shufflein.R
\name{shufflein}
\alias{shufflein}
\title{Return the In Shuffle of a vector}
\usage{
shufflein(v, n = 1)
}
\arguments{
\item{v}{Input Vector}

\item{n}{Number of Shuffles}
}
\value{
Shuffled Vector
}
\description{
Return the In Shuffle of a vector
}
\examples{
shufflein(0:9,1)
}
