% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffleout.R
\name{shuffleout}
\alias{shuffleout}
\title{Return the Out Shuffle of a vector}
\usage{
shuffleout(v, n = 1)
}
\arguments{
\item{v}{Input Vector}

\item{n}{Number of Shuffles}
}
\value{
Shuffled Vector
}
\description{
Return the Out Shuffle of a vector
}
\examples{
shuffleout(0:9,1)
}
