% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableprop.R
\name{tableprop}
\alias{tableprop}
\title{Display Count and Proportions of discrete Variable}
\usage{
tableprop(v)
}
\arguments{
\item{v}{Vector of discrete variable}
}
\description{
Display the table count and proportions of variable discrete values
}
