% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tavily.R
\name{tavily}
\alias{tavily}
\title{Submit a search to Tavily Service}
\usage{
tavily(prompt, max_results = 1)
}
\arguments{
\item{prompt}{Text prompt}
}
\description{
Submit a search to Tavily Service
}
\examples{
\dontrun{
tavily("What is Langgraph?")
}
}
