% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminal.R
\name{terminal}
\alias{terminal}
\title{Send command to new terminal}
\usage{
terminal(command)
}
\arguments{
\item{command}{Send command to terminal}
}
\description{
Send command to new terminal
}
\examples{
\dontrun{
terminal("echo 'Hello World'")
}
}
