% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truthtable.R
\name{truthtable}
\alias{truthtable}
\title{Return The Truth Table of the Input Function as a Data Frame}
\usage{
truthtable(f)
}
\arguments{
\item{f}{A Logical Function - A Function that returns Truth or False}
}
\value{
The Truth Table of the Input Function as a Data Frame
}
\description{
Return The Truth Table of the Input Function as a Data Frame
}
\examples{
truthtable( function(A,B,C,D) A && B && (!C) && (!D) )
}
