% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagegrid.R
\name{imagegrid}
\alias{imagegrid}
\title{Display the images in a folder as a grid using the magick package}
\usage{
imagegrid(folder, ncol = 0, nrow = 0)
}
\arguments{
\item{folder}{Images folder}

\item{ncol}{Optional grid number of columns}

\item{nrow}{Optional grid number of rows}
}
\description{
Display the images in a folder as a grid using the magick package
}
\examples{
\dontrun{
imagegrid("P:/DISKD/Youtube R Programming/LOGOS")
}
}
