% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagerenamer.R
\name{imagerenamer}
\alias{imagerenamer}
\title{Mass Rename and extension fix for images in a folder}
\usage{
imagerenamer(folder = getwd(), image_prefix)
}
\arguments{
\item{folder}{Optional Images Folder. Defaults to Current Working Directory}

\item{image_prefix}{Optional Image Prefix to be used on new names}
}
\description{
Mass Rename and extension fix for images in a folder
}
\examples{
\dontrun{
imagerenamer("P:/DISKD/Wallpapers","IMG")
}
}
