% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printdataframe.R
\name{printdataframe}
\alias{printdataframe}
\title{Print Data Frame as HTML Table}
\usage{
printdataframe(df)
}
\arguments{
\item{df}{Data Frame to be printed as HTML Table}
}
\description{
Print Data Frame as HTML Table
}
\examples{
\dontrun{
printdataframe(iris)
}
}
